+++ client/src/components/Search/style.css
