+++ client/src/views/Find/Find.tsx
@@ -1,13 +1,14 @@
 import { withRouter } from 'react-router';
 
 import UserCard from '../../components/Users/UserCards'
-
+// import SearchStatus from '../../components/Search/SearchStatus'
+import './style.css'
 const Find: React.FC = (): JSX.Element => {
 
 
     return (
-        <div>
-     
+        <div className="paddingfix" >
+            {/* <SearchStatus /> */}
             <UserCard />
         
         </div>
