+++ client/src/views/Find/style.css
@@ -1,0 +1,11 @@
+.bg{
+    /* background-color: blueviolet; */
+    background-color: rgba(255, 255, 255, 0.7);
+    padding-top: 5rem;
+    align-items: center;
+    text-align: center;
+} 
+
+.paddingfix {
+    padding-top: 100px;
+}
