+++ client/src/views/Landing/loggedInLanding.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Switch } from "react-router-dom";
 import Find from '../Find/Find';
 import TempProfile from '../CreateProfile/TempProfile';
-import UserProfile from '../../components/UserProfile/UserProfile'
+import UserProfile from '../UserProfile/UserProfile'
 import CreateProfile from '../CreateProfile/CreateProfile';
 import ProtectedRoute from '../../auth/protected-route';
